Moved all notes related components inside notes directory.
