+++ Frontend/src/components/notes/Addnote.jsx
@@ -1,5 +1,5 @@
 import { useState, useContext } from "react";
-import noteContext from "../context/Notes/noteContext";
+import noteContext from "../../context/Notes/noteContext";
 const Addnote = () => {
   /* ----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------- */
   /* ----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------- */
